주요 함수 분리/ 모듈화
