+++ displaymap.h
@@ -1,0 +1,15 @@
+#ifndef DISPLAYMAP_H
+#define DISPLAYMAP_H
+
+#include "utility.h"
+#include "initmap.h"
+#include "log.h"
+#include "player.h"
+#include "enemy.h"
+
+
+void displayMap();
+void displayBattleScreen();
+
+
+#endif // DISPLAYMAP_H
