+++ initmap.h
@@ -1,0 +1,9 @@
+#ifndef INITMAP_H
+#define INITMAP_H
+
+#include "utility.h"
+
+void initializeMap();
+void initializeMapBattle();
+
+#endif // INITMAP_H
