+++ startscreen.cpp
@@ -1,0 +1,59 @@
+#include "startscreen.h"
+#include <stdio.h>
+
+void start_screen() {	//시작화면
+	for (int i = 0; i < (MAP_HEIGHT - 16) / 2; i++)
+		printf("\n");
+
+	for (int j = 0; j < (MAP_WIDTH - 57) / 2; j++)
+		printf(" ");
+	printf("=========================================================\n");
+	for (int j = 0; j < (MAP_WIDTH - 57) / 2; j++)
+		printf(" ");
+	printf("||                                                   ||\n");
+	for (int j = 0; j < (MAP_WIDTH - 57) / 2; j++)
+		printf(" ");
+	printf("||                임  진  투  쟁  기                 ||\n");
+	for (int j = 0; j < (MAP_WIDTH - 57) / 2; j++)
+		printf(" ");
+	printf("||                                                   ||\n");
+	for (int j = 0; j < (MAP_WIDTH - 57) / 2; j++)
+		printf(" ");
+	printf("=========================================================\n");
+	for (int j = 0; j < (MAP_WIDTH - 57) / 2; j++)
+		printf(" ");
+	printf("         Text-based RPG Game Set in the Joseon Dynasty\n\n");
+	for (int j = 0; j < (MAP_WIDTH - 57) / 2; j++)
+		printf(" ");
+	printf("           Press [Enter] to start your adventure!\n");
+	for (int j = 0; j < (MAP_WIDTH - 57) / 2; j++)
+		printf(" ");
+	printf("=========================================================\n\n");
+	for (int j = 0; j < (MAP_WIDTH - 57) / 2; j++)
+		printf(" ");
+	printf("                        |         |  |\n");
+	for (int j = 0; j < (MAP_WIDTH - 57) / 2; j++)
+		printf(" ");
+	printf("                      ㅡㅡㅡ \\/  ㅡㅡㅡ\n");
+	for (int j = 0; j < (MAP_WIDTH - 57) / 2; j++)
+		printf(" ");
+	printf("                        O    /\u005C    O\n");
+	for (int j = 0; j < (MAP_WIDTH - 57) / 2; j++)
+		printf(" ");
+	printf("                       /|\\  /  \\  /|\\\n");
+	for (int j = 0; j < (MAP_WIDTH - 57) / 2; j++)
+		printf(" ");
+	printf("                      / | \\/    \\/ | \\\n");
+	for (int j = 0; j < (MAP_WIDTH - 57) / 2; j++)
+		printf(" ");
+	printf("                       / \\        / \\\n");
+	for (int j = 0; j < (MAP_WIDTH - 57) / 2; j++)
+		printf(" ");
+	printf("                      /   \\      /   \\\n\n");
+	for (int j = 0; j < (MAP_WIDTH - 57) / 2; j++)
+		printf(" ");
+	printf("=========================================================\n");
+
+	getchar();  // 사용자로부터 Enter 입력 대기
+
+}
