+++ startscreen.h
@@ -1,0 +1,8 @@
+#ifndef STARTSCREEN_H
+#define STARTSCREEN_H
+
+#include "utility.h"
+
+void start_screen();
+
+#endif // ENEMY_H
