+++ utility.cpp
@@ -1,0 +1,24 @@
+#include "utility.h"
+
+//전투맵, 일반맵
+char map[MAP_HEIGHT][MAP_WIDTH + PANEL_WIDTH];
+char mapBattle[MAP_HEIGHT][MAP_WIDTH + PANEL_WIDTH];
+int Situation = 0;
+
+// 커서를 특정 위치로 이동시키는 함수
+void setCursorPosition(int x, int y)
+{
+	COORD pos = { x, y };
+	HANDLE hConsoleOut = GetStdHandle(STD_OUTPUT_HANDLE);
+	SetConsoleCursorPosition(hConsoleOut, pos);
+}
+
+//커서를 지우는 함수
+void eraseCursor()
+{
+	HANDLE hConsoleOut = GetStdHandle(STD_OUTPUT_HANDLE);
+	CONSOLE_CURSOR_INFO  curCursorInfo;
+	GetConsoleCursorInfo(hConsoleOut, &curCursorInfo);
+	curCursorInfo.bVisible = 0;
+	SetConsoleCursorInfo(hConsoleOut, &curCursorInfo);
+}
